+++ tests/test_oauth2_resource_owner_password_okta.py
@@ -13,15 +13,17 @@ def test_oauth2_password_credentials_flow_uses_provided_session(
 ):
     session = requests.Session()
     session.headers.update({"x-test": "Test value"})
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token",
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
         username="test_user",
         password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
         session=session,
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -34,78 +36,29 @@ def test_oauth2_password_credentials_flow_uses_provided_session(
         get_header(responses, auth).get("Authorization")
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
-    request = get_request(responses, "http://provide_access_token/")
-    assert request.body == "grant_type=password&username=test_user&password=test_pwd"
+    request = get_request(
+        responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+    )
+    assert (
+        request.body
+        == "grant_type=password&username=test_user&password=test_pwd&scope=openid"
+    )
     assert request.headers["x-test"] == "Test value"
 
 
 def test_oauth2_password_credentials_flow_token_is_sent_in_authorization_header_by_default(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
-    )
-    responses.add(
-        responses.POST,
-        "http://provide_access_token",
-        json={
-            "access_token": "2YotnFZFEjr1zCsicMWpAA",
-            "token_type": "example",
-            "expires_in": 3600,
-            "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
-            "example_parameter": "example_value",
-        },
-    )
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Bearer 2YotnFZFEjr1zCsicMWpAA"
-    )
-    assert (
-        get_request(responses, "http://provide_access_token/").body
-        == "grant_type=password&username=test_user&password=test_pwd"
-    )
-
-
-def test_oauth2_password_credentials_flow_does_not_authenticate_by_default(
-    token_cache, responses: RequestsMock
-):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
-    )
-    responses.add(
-        responses.POST,
-        "http://provide_access_token",
-        json={
-            "access_token": "2YotnFZFEjr1zCsicMWpAA",
-            "token_type": "example",
-            "expires_in": 3600,
-            "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
-            "example_parameter": "example_value",
-        },
-    )
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Bearer 2YotnFZFEjr1zCsicMWpAA"
-    )
-    token_request = get_request(responses, "http://provide_access_token/")
-    assert (
-        token_request.body == "grant_type=password&username=test_user&password=test_pwd"
-    )
-    assert "Authorization" not in token_request.headers
-
-
-def test_oauth2_password_credentials_flow_authentication(
-    token_cache, responses: RequestsMock
-):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token",
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
         username="test_user",
         password="test_pwd",
-        session_auth=("test_user2", "test_pwd2"),
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -118,9 +71,12 @@ def test_oauth2_password_credentials_flow_authentication(
         get_header(responses, auth).get("Authorization")
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
-    token_request = get_request(responses, "http://provide_access_token/")
+    token_request = get_request(
+        responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+    )
     assert (
-        token_request.body == "grant_type=password&username=test_user&password=test_pwd"
+        token_request.body
+        == "grant_type=password&username=test_user&password=test_pwd&scope=openid"
     )
     assert (
         "Basic dGVzdF91c2VyMjp0ZXN0X3B3ZDI=" == token_request.headers["Authorization"]
@@ -130,19 +86,23 @@ def test_oauth2_password_credentials_flow_authentication(
 def test_oauth2_password_credentials_flow_token_is_expired_after_30_seconds_by_default(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     # Add a token that expires in 29 seconds, so should be considered as expired when issuing the request
     token_cache._add_token(
-        key="db2be9203dd2718c7285319dde1270056808482fbf7fffa6a9362d092d1cf799b393dd15140ea13e4d76d1603e56390a6222ff7063736a1b686d317706b2c001",
+        key="bdc39831ac59c0f65d36761e9b65656ae76223f2284c393a6e93fe4e09a2c0002e2638bbe02db2cc62928a2357be5e2e93b9fa4ac68729f4d28da180caae912a",
         token="2YotnFZFEjr1zCsicMWpAA",
         expiry=requests_auth.oauth2_tokens._to_expiry(expires_in=29),
     )
     # Meaning a new one will be requested
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -156,23 +116,27 @@ def test_oauth2_password_credentials_flow_token_is_expired_after_30_seconds_by_d
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
     assert (
-        get_request(responses, "http://provide_access_token/").body
-        == "grant_type=password&username=test_user&password=test_pwd"
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
+        == "grant_type=password&username=test_user&password=test_pwd&scope=openid"
     )
 
 
 def test_oauth2_password_credentials_flow_token_custom_expiry(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token",
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
         username="test_user",
         password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
         early_expiry=28,
     )
     # Add a token that expires in 29 seconds, so should be considered as not expired when issuing the request
     token_cache._add_token(
-        key="db2be9203dd2718c7285319dde1270056808482fbf7fffa6a9362d092d1cf799b393dd15140ea13e4d76d1603e56390a6222ff7063736a1b686d317706b2c001",
+        key="bdc39831ac59c0f65d36761e9b65656ae76223f2284c393a6e93fe4e09a2c0002e2638bbe02db2cc62928a2357be5e2e93b9fa4ac68729f4d28da180caae912a",
         token="2YotnFZFEjr1zCsicMWpAA",
         expiry=requests_auth.oauth2_tokens._to_expiry(expires_in=29),
     )
@@ -183,12 +147,16 @@ def test_oauth2_password_credentials_flow_token_custom_expiry(
 
 
 def test_expires_in_sent_as_str(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -202,19 +170,25 @@ def test_expires_in_sent_as_str(token_cache, responses: RequestsMock):
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
     assert (
-        get_request(responses, "http://provide_access_token/").body
-        == "grant_type=password&username=test_user&password=test_pwd"
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
+        == "grant_type=password&username=test_user&password=test_pwd&scope=openid"
     )
 
 
 def test_refresh_token(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     # response for password grant
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -228,6 +202,7 @@ def test_refresh_token(token_cache, responses: RequestsMock):
                     "grant_type": "password",
                     "username": "test_user",
                     "password": "test_pwd",
+                    "scope": "openid",
                 }
             )
         ],
@@ -238,14 +213,16 @@ def test_refresh_token(token_cache, responses: RequestsMock):
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
     assert (
-        get_request(responses, "http://provide_access_token/").body
-        == "grant_type=password&username=test_user&password=test_pwd"
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
+        == "grant_type=password&username=test_user&password=test_pwd&scope=openid"
     )
 
     # response for refresh token grant
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "rVR7Syg5bjZtZYjbZIW",
             "token_type": "example",
@@ -258,6 +235,7 @@ def test_refresh_token(token_cache, responses: RequestsMock):
                 {
                     "grant_type": "refresh_token",
                     "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
+                    "scope": "openid",
                 }
             )
         ],
@@ -266,19 +244,25 @@ def test_refresh_token(token_cache, responses: RequestsMock):
     response = requests.get("http://authorized_only", auth=auth)
     assert response.request.headers.get("Authorization") == "Bearer rVR7Syg5bjZtZYjbZIW"
     assert (
-        get_request(responses, "http://provide_access_token/").body
-        == "grant_type=refresh_token&refresh_token=tGzv3JOkF0XG5Qx2TlKWIA"
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
+        == "grant_type=refresh_token&scope=openid&refresh_token=tGzv3JOkF0XG5Qx2TlKWIA"
     )
 
 
 def test_refresh_token_invalid(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     # response for password grant
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -292,6 +276,7 @@ def test_refresh_token_invalid(token_cache, responses: RequestsMock):
                     "grant_type": "password",
                     "username": "test_user",
                     "password": "test_pwd",
+                    "scope": "openid",
                 }
             )
         ],
@@ -302,14 +287,16 @@ def test_refresh_token_invalid(token_cache, responses: RequestsMock):
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
     assert (
-        get_request(responses, "http://provide_access_token/").body
-        == "grant_type=password&username=test_user&password=test_pwd"
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
+        == "grant_type=password&username=test_user&password=test_pwd&scope=openid"
     )
 
     # response for refresh token grant
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "invalid_request"},
         status=400,
         match=[
@@ -317,6 +304,7 @@ def test_refresh_token_invalid(token_cache, responses: RequestsMock):
                 {
                     "grant_type": "refresh_token",
                     "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
+                    "scope": "openid",
                 }
             )
         ],
@@ -330,13 +318,17 @@ def test_refresh_token_invalid(token_cache, responses: RequestsMock):
 
 
 def test_refresh_token_access_token_not_expired(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     # response for password grant
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -350,6 +342,7 @@ def test_refresh_token_access_token_not_expired(token_cache, responses: Requests
                     "grant_type": "password",
                     "username": "test_user",
                     "password": "test_pwd",
+                    "scope": "openid",
                 }
             )
         ],
@@ -360,8 +353,10 @@ def test_refresh_token_access_token_not_expired(token_cache, responses: Requests
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
     assert (
-        get_request(responses, "http://provide_access_token/").body
-        == "grant_type=password&username=test_user&password=test_pwd"
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
+        == "grant_type=password&username=test_user&password=test_pwd&scope=openid"
     )
 
     # expect Bearer token to remain the same
@@ -372,15 +367,17 @@ def test_refresh_token_access_token_not_expired(token_cache, responses: Requests
 
 
 def test_scope_is_sent_as_is_when_provided_as_str(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token",
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
         username="test_user",
         password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
         scope="my_scope+my_other_scope",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -394,7 +391,9 @@ def test_scope_is_sent_as_is_when_provided_as_str(token_cache, responses: Reques
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
     assert (
-        get_request(responses, "http://provide_access_token/").body
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
         == "grant_type=password&username=test_user&password=test_pwd&scope=my_scope%2Bmy_other_scope"
     )
 
@@ -402,15 +401,17 @@ def test_scope_is_sent_as_is_when_provided_as_str(token_cache, responses: Reques
 def test_scope_is_sent_as_str_when_provided_as_list(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token",
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
         username="test_user",
         password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
         scope=["my_scope", "my_other_scope"],
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -424,17 +425,26 @@ def test_scope_is_sent_as_str_when_provided_as_list(
         == "Bearer 2YotnFZFEjr1zCsicMWpAA"
     )
     assert (
-        get_request(responses, "http://provide_access_token/").body
+        get_request(
+            responses, "https://testserver.okta-emea.com/oauth2/default/v1/token"
+        ).body
         == "grant_type=password&username=test_user&password=test_pwd&scope=my_scope+my_other_scope"
     )
 
 
 def test_with_invalid_grant_request_no_json(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
-        responses.POST, "http://provide_access_token", body="failure", status=400
+        responses.POST,
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
+        body="failure",
+        status=400,
     )
     with pytest.raises(requests_auth.InvalidGrantRequest) as exception_info:
         requests.get("http://authorized_only", auth=auth)
@@ -444,12 +454,16 @@ def test_with_invalid_grant_request_no_json(token_cache, responses: RequestsMock
 def test_with_invalid_grant_request_invalid_request_error(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "invalid_request"},
         status=400,
     )
@@ -467,12 +481,16 @@ def test_with_invalid_grant_request_invalid_request_error(
 def test_with_invalid_grant_request_invalid_request_error_and_error_description(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "invalid_request", "error_description": "desc of the error"},
         status=400,
     )
@@ -484,12 +502,16 @@ def test_with_invalid_grant_request_invalid_request_error_and_error_description(
 def test_with_invalid_grant_request_invalid_request_error_and_error_description_and_uri(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "error": "invalid_request",
             "error_description": "desc of the error",
@@ -508,12 +530,16 @@ def test_with_invalid_grant_request_invalid_request_error_and_error_description_
 def test_with_invalid_grant_request_invalid_request_error_and_error_description_and_uri_and_other_fields(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "error": "invalid_request",
             "error_description": "desc of the error",
@@ -531,12 +557,16 @@ def test_with_invalid_grant_request_invalid_request_error_and_error_description_
 
 
 def test_with_invalid_grant_request_without_error(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"other": "other info"},
         status=400,
     )
@@ -548,12 +578,16 @@ def test_with_invalid_grant_request_without_error(token_cache, responses: Reques
 def test_with_invalid_grant_request_invalid_client_error(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "invalid_client"},
         status=400,
     )
@@ -575,12 +609,16 @@ def test_with_invalid_grant_request_invalid_client_error(
 def test_with_invalid_grant_request_invalid_grant_error(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "invalid_grant"},
         status=400,
     )
@@ -598,12 +636,16 @@ def test_with_invalid_grant_request_invalid_grant_error(
 def test_with_invalid_grant_request_unauthorized_client_error(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "unauthorized_client"},
         status=400,
     )
@@ -619,12 +661,16 @@ def test_with_invalid_grant_request_unauthorized_client_error(
 def test_with_invalid_grant_request_unsupported_grant_type_error(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "unsupported_grant_type"},
         status=400,
     )
@@ -640,12 +686,16 @@ def test_with_invalid_grant_request_unsupported_grant_type_error(
 def test_with_invalid_grant_request_invalid_scope_error(
     token_cache, responses: RequestsMock
 ):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token", username="test_user", password="test_pwd"
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
+        username="test_user",
+        password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={"error": "invalid_scope"},
         status=400,
     )
@@ -659,15 +709,17 @@ def test_with_invalid_grant_request_invalid_scope_error(
 
 
 def test_without_expected_token(token_cache, responses: RequestsMock):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token",
+    auth = requests_auth.OktaResourceOwnerPasswordCredentials(
+        "testserver.okta-emea.com",
         username="test_user",
         password="test_pwd",
+        client_id="test_user2",
+        client_secret="test_pwd2",
         token_field_name="not_provided",
     )
     responses.add(
         responses.POST,
-        "http://provide_access_token",
+        "https://testserver.okta-emea.com/oauth2/default/v1/token",
         json={
             "access_token": "2YotnFZFEjr1zCsicMWpAA",
             "token_type": "example",
@@ -684,33 +736,74 @@ def test_without_expected_token(token_cache, responses: RequestsMock):
     )
 
 
-def test_token_url_is_mandatory():
+def test_instance_is_mandatory():
     with pytest.raises(Exception) as exception_info:
-        requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-            "", "test_user", "test_pwd"
+        requests_auth.OktaResourceOwnerPasswordCredentials(
+            "",
+            username="test_user",
+            password="test_pwd",
+            client_id="test_user2",
+            client_secret="test_pwd2",
         )
-    assert str(exception_info.value) == "Token URL is mandatory."
+    assert str(exception_info.value) == "Instance is mandatory."
 
 
 def test_user_name_is_mandatory():
     with pytest.raises(Exception) as exception_info:
-        requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-            "http://test_url", "", "test_pwd"
+        requests_auth.OktaResourceOwnerPasswordCredentials(
+            "testserver.okta-emea.com",
+            username="",
+            password="test_pwd",
+            client_id="test_user2",
+            client_secret="test_pwd2",
         )
     assert str(exception_info.value) == "User name is mandatory."
 
 
 def test_password_is_mandatory():
     with pytest.raises(Exception) as exception_info:
-        requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-            "http://test_url", "test_user", ""
+        requests_auth.OktaResourceOwnerPasswordCredentials(
+            "testserver.okta-emea.com",
+            username="test_user",
+            password="",
+            client_id="test_user2",
+            client_secret="test_pwd2",
         )
     assert str(exception_info.value) == "Password is mandatory."
 
 
+def test_client_id_is_mandatory():
+    with pytest.raises(Exception) as exception_info:
+        requests_auth.OktaResourceOwnerPasswordCredentials(
+            "testserver.okta-emea.com",
+            username="test_user",
+            password="test_pwd",
+            client_id="",
+            client_secret="test_pwd2",
+        )
+    assert str(exception_info.value) == "Client ID is mandatory."
+
+
+def test_client_secret_is_mandatory():
+    with pytest.raises(Exception) as exception_info:
+        requests_auth.OktaResourceOwnerPasswordCredentials(
+            "testserver.okta-emea.com",
+            username="test_user",
+            password="test_pwd",
+            client_id="test_user2",
+            client_secret="",
+        )
+    assert str(exception_info.value) == "Client secret is mandatory."
+
+
 def test_header_value_must_contains_token():
     with pytest.raises(Exception) as exception_info:
-        requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-            "http://test_url", "test_user", "test_pwd", header_value="Bearer token"
+        requests_auth.OktaResourceOwnerPasswordCredentials(
+            "testserver.okta-emea.com",
+            username="test_user",
+            password="test_pwd",
+            client_id="test_user2",
+            client_secret="test_pwd2",
+            header_value="Bearer token",
         )
     assert str(exception_info.value) == "header_value parameter must contains {token}."